Move the Import BIND file button under the DNS Actions menu
